pyproject.toml
[project]
name = "reasoningbank"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.5.0",
    "langchain-core>=1.2.13",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "matplotlib>=3.10.8",
    "mypy>=1.19.1",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.1",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.2",
    "tqdm>=4.67.3",
    "wandb>=0.25.0",
]
